Turn solution into helper function and apply solution to same issue on Query Report sorting
