+++ frontend/pages/queries/edit/components/QueryResults/QueryResults.tsx
@@ -5,16 +5,13 @@ import classnames from "classnames";
 import FileSaver from "file-saver";
 import { QueryContext } from "context/query";
 import { useDebouncedCallback } from "use-debounce";
-import { find } from "lodash";
 
 import {
   generateCSVFilename,
   generateCSVQueryResults,
 } from "utilities/generate_csv";
-import { osqueryTables } from "utilities/osquery_tables";
 import { ICampaign, ICampaignError } from "interfaces/campaign";
 import { ITarget } from "interfaces/target";
-import { IQueryTableColumn } from "interfaces/osquery_table";
 
 import Button from "components/buttons/Button";
 import Icon from "components/Icon/Icon";
@@ -25,9 +22,9 @@ import QueryResultsHeading from "components/queries/queryResults/QueryResultsHea
 import AwaitingResults from "components/queries/queryResults/AwaitingResults";
 import InfoBanner from "components/InfoBanner";
 import CustomLink from "components/CustomLink";
-import { checkTable } from "utilities/sql_tools";
 
 import generateColumnConfigsFromRows from "./QueryResultsTableConfig";
+import { getTableColumnsFromSql } from "utilities/helpers";
 
 interface IQueryResultsProps {
   campaign: ICampaign;
@@ -77,9 +74,6 @@ const QueryResults = ({
   const [queryResultsForTableRender, setQueryResultsForTableRender] = useState(
     queryResults
   );
-  const [osqueryTableColumns, setOsqueryTableColumns] = useState<
-    IQueryTableColumn[] | []
-  >([]);
 
   // immediately reset results
   const onRunAgain = useCallback(() => {
@@ -98,32 +92,20 @@ const QueryResults = ({
     debounceQueryResults(queryResults);
   }, [queryResults, debounceQueryResults]);
 
-  // Set table/s columns from SQL
-  useEffect(() => {
-    const tableNames =
-      (lastEditedQueryBody && checkTable(lastEditedQueryBody).tables) || [];
-
-    let columns: IQueryTableColumn[] | [] = [];
-    tableNames.forEach((tableName: string) => {
-      const tableColumns =
-        find(osqueryTables, { name: tableName })?.columns || [];
-      columns = [...columns, ...tableColumns];
-    });
-    setOsqueryTableColumns(columns);
-  }, [lastEditedQueryBody]);
-
   useEffect(() => {
     if (queryResults && queryResults.length > 0) {
+      const tableColumns = getTableColumnsFromSql(lastEditedQueryBody);
+
       const newResultsColumnConfigs = generateColumnConfigsFromRows(
         queryResults,
-        osqueryTableColumns
+        tableColumns
       );
       // Update tableHeaders if new headers are found
       if (newResultsColumnConfigs !== resultsColumnConfigs) {
         setResultsColumnConfigs(newResultsColumnConfigs);
       }
     }
-  }, [queryResults]); // Cannot use tableHeaders as it will cause infinite loop with setTableHeaders
+  }, [queryResults, lastEditedQueryBody]); // Cannot use tableHeaders as it will cause infinite loop with setTableHeaders
 
   useEffect(() => {
     if (errorColumnConfigs?.length === 0 && !!errors?.length) {
